+++ DrivingLicenseAppAPI/src/main/java/csapat/DrivingLicenseAppAPI/dto/InstructorUpdate.java
@@ -2,10 +2,10 @@ package csapat.DrivingLicenseAppAPI.dto;
 
 public record InstructorUpdate(
         String promoText,
-        Integer vehicleId,
+        Long vehicleId,
         String vehicleName,
         String licensePlate,
-        Integer fuelTypeId,
-        Integer vehicleTypeId
+        Long fuelTypeId,
+        Long vehicleTypeId
 ) {
 }
